Subset the cached timesteps before widening them to float32
Measured the unpacked read path: 9.6 ms per sample, of which the npz
branch was 7.1 ms - and that was not I/O. The loader converted the whole
K=7 grid from float16 to float32 and then discarded whatever _pick did
not select, so most of a 3.5 MiB expansion was thrown away. Subsetting
first makes the conversion proportional to K.

This also settles whether the pack is worth running. At batch 32 the
unpacked loader needs 308 ms of CPU, about 77 ms of wall across 4
workers, against a ~658 ms stage-1 step - already 8x ahead of the GPU,
so the cost is entirely hidden behind compute. Packing takes that to
28 ms of something that is not on the critical path. pack_offline_gt.py
stays for the case where training logs show the loader stalling, but it
is not part of the normal flow, and the cache should not be persisted
twice.

+++ HermiteFlow-Code/src/datasets/x4k_cached.py
@@ -272,13 +272,21 @@ class X4KCachedGT(Dataset):
             path = self.samples[index]
             prefix = path[: -len("_flow.npz")]
             with np.load(path) as data:
-                phi_all = data["flow_0_t"].astype(np.float32)  # (Kc, 2, H, W)
-                psi_all = data["flow_1_t"].astype(np.float32)
+                phi_all = data["flow_0_t"]          # (Kc, 2, H, W) float16
+                psi_all = data["flow_1_t"]
                 t_all = data["t"].astype(np.float32)
 
-            keep = self._pick(len(t_all)) if self.if_aug else self._eval_pick(len(t_all))
-            phi = [phi_all[i] for i in keep]
-            psi = [psi_all[i] for i in keep]
+                # Subset BEFORE widening to float32. Converting the whole
+                # grid and then discarding what _pick did not choose was
+                # the dominant cost in this branch - 7.1 ms of a 9.6 ms
+                # sample, nearly all of it the fp16->fp32 expansion of
+                # 3.5 MiB, of which K/Kc is actually used.
+                keep = (
+                    self._pick(len(t_all)) if self.if_aug
+                    else self._eval_pick(len(t_all))
+                )
+                phi = [phi_all[i].astype(np.float32) for i in keep]
+                psi = [psi_all[i].astype(np.float32) for i in keep]
             times = [float(t_all[i]) for i in keep]
 
             # cv2 reads BGR; the cache was written from RGB, reverse back.
